refactor: DailyRecPanel 차트 섹션을 컴포넌트로 분리 #60
